Line, Oval, Shape now always paint themselves
This used to be optional, in case an object was actually part of
a Shape (in which the Shape was in charge of painting). As of
f7d32f576451c92835a2fea5fa93d6a97113569a, Shapes are no longer
composites of other objects, so this option is no longer needed.