Add Rake task to run all working samples
Usage: rake samples

This task reads the list of working samples from samples/README,
shuffles the list, and executes each sample. When you quit the current
sample, the next one spins up. You wouldn't want to run this all the
time (it takes 7 or 8 minutes on my machine at the moment), but it's
a nice sanity check.