Remove Shoes::Color.to_native
Instead, add a factory: Shoes::Swt::Color.create(color) that
creates as Shoes::Swt::Color from a Shoes::Color.

Benefit: Gracefully handles the case where we have a nil for the
DSL object's fill or stroke. In this case, the factory creates a
NullColor (a null object). This way, we don't have to check for nil
before we do painting--the NullColor's paint operations are no-ops.

This commit also removes the Shoes::Swt::Color component methods
(#red, #green, #blue). There's really no reason to ask the backend
for these values.