Refactorings in dsl.rb, and edit_box 'with_text'.
This is a partial cleanup in dsl.rb, passing the app
in a consistent way to the elements, and putting that
boilerplate stuff in one place.

It also adds missing behavior for edit_box, which can
now be called with an intial text as a first parameter.

We also added some specs for DSL methods that were still
untested. edit_box and image.

Shoes on! @bjoska and @arnebrasseur :metal: