Implement Shoes.setup to print a warning message
Shoes 3 Had a way of specifying gems a Shoes app depends on, like so

    Shoes.setup do
      gem 'nokogiri ~>1.6.0'
    end

This was relevant then because Shoes embedded its own Ruby, in this case it
would take care of installing the gem in the right place and loading it in
the app.

In Shoes 4 we will no longer do this, the user can use the standard tools
(gem install, or bundler on top of that) and require the needed files in
the Shoes app.

To ease the upgrade path for people coming from Shoes 3 (or relying on outdated
documentation) we print a warning telling them what to do. This is what it looks
like.

    WARN: The Shoes.setup method is no longer used, you need to install gems yourself.
    WARN: To use the 'bluecloth' gem, install it with 'gem install bluecloth --version "=2.0.6"', and put 'require "bluecloth"' at the top of your Shoes program.
    WARN: To use the 'metaid' gem, install it with 'gem install metaid ', and put 'require "metaid"' at the top of your Shoes program.