Fix for button sizing #452
Turns out that the placement of where the dimensioning was in relation to
the yield(@real) was important. This is because the yield sets the button text
and that in turn influences its width. Setting the size before the yield just
gets whatever button size is out of the box with an empty string.