Partial fix for TextBlock flowing
Turns out that while we were grabbing the calculated height of a text block
line from the SWT side, we were not doing the same properly from the width,
and were tromping the value in when Shoes::TextBlock#set_size was called.

This doesn't address multi-line text blocks and flowing them properly. That
will still take quite a bit of revamping, but this gets smaller blocks acting
like they did in Shoes 3.