Fixing and refactoring assignment of absolute_top
When we are looking to set the absolute_top for advancing the current_position,
this has to take into account 1) the height of the second layout 2) the current
y position, 3) the height of the last line of text (which our next positioning
needs to align with the top, not the bottom of), and 4) the spacing on the
layout.

Whew. With all that right, things line up. :fingers_crossed: