Refactoring and removing unneeded cursor logic
Some of the logic flow of positioning the cursor was relying too much on the
order to determine if we were in a single layout. Turns out, that wasn't
necessary given the later checks for exceeding the full text length, so cleaned
that up.

Also extracted some more precisely named methods for how things get positioned
to help with readability.