Adding helpers for test readability
The text fitting tests get kind of heavy on all the coordinates and exact
values, but without them they aren't very meaningful.

This change introduces more helpers to give names to clusters of specific
setup and checks we do. For instance, using a hash so we can give numbers
useful naming at the call sites, and helpers for asserting positioning after
the fact more expressively.