Pass options through on DSL link
Calls to the link method on the DSL weren't passing along options properly when
presented as a hash in the last argument of the call. These were instead
interpretted as a final text element (see #588 for an example).

Shoes::DSL#pop_style already exists to handle exactly this, just needed to get
called correctly from the link DSL method.