Support calling Link#click to change block
Originally the Shoes::Link class would only support setting the block at
instantiation. Shoes3, though, provided the #click method to override the
existing block on a link.

This commit restores that. Along with plumbing it into the Link class, it also
had to change the Shoes::Swt::Link class's wiring up to the click listener so
it wouldn't capture the original block but instead delegate to the link itself.