Promote CommonMethods DSL classes to top level
Older Rubies that Shoes3 ran on had scoping differences that meant that DSL
classes could be referred to unqualified (i.e. Link instead of Shoes::Link).

To retain compatibility with that older code (and also keep the Shoes API
simpler), we promote any DSL class that gets Shoes::CommonMethods included in
it to the top-level.