Fix for layout fitting when no space on first line
Previously we've made the assumption that there's enough room on the first line
when fitting text to put something there. Turns out we can have too little
space, and in those cases we'd end up crashing on some negative-width setting.

Instead, the right behavior in these cases is to provide an empty first layout,
and then proceed to put the actual text in a typically flowed second layout.

Resolves #594.