Fix for hang during painting
We had an odd problem where opening an alert dialog right after adding
an element would cause hangs in the app. This was a result of the
element not having been placed yet, and the painter for it crashing in a
spot that SWT wasn't good about reporting the error back.

This commit now provides a check for the object to make sure that it's
positioned, and also has a rescue clause to keep errors from escaping.