Bump to parent widths if positioned outside
If a text block is forced by the app code to position outside of it's
container, our fitting logic wasn't really happy. We now will search
outward to find a fitting width in those rare cases, and also bail with
a stand-in value if we can't find somewhere to go.