Centering bumps to new line
For now to keep things simple and reasonable, if a text block is
centered, then we bump it down to its own line and carry on from there.

This is accomplished via a similar technique to the empty layout we use
when there's not enough space left on the starting line for any real
text with normal text blocks.