Fix for offsetting of second text segments
Was noted on issue #717 that we had a problem with the text blocks on
the later lines not quite lining up when things wrapped.

Turned out this was because when we asked the first layout its height to
determine where the second layout should begin vertically, that value
included the spacing which contributed +4 pixels that we didn't actually
want. Removing that we I found we still needed one of those pixels,
so we adjust the second segment's starting top accordingly.

Note there's still a slight offset + background issue, but the text now
lines up between the segments properly.