Actually run all the specs :(
We were missing the ** in our file lookup for SWT specs which meant that
specs in subdirectories (of which there weren't many) weren't running.

Fixing that I found that the painter spec for SWT wasn't running
properly, and also that because of naming we were requiring the shared
example for it too often, generating a warning.