Spec fixes to support adjust_current_position
Shoes::Swt::TextBlock used to set absolute_top since that would get
passed along to the current_position.y by the slot. That was a misuse of
the side-effect, so now we provide an explicit mechanism to allow
altering the current_position from an element.

This commit updats the specs to align with that reality.