Initial support for shoes backend executables
This introduces new functionality for chosing a shoes backend.  When a
backend is chosen, we write to `shoes-backend` in the bin directory,
with a generated command from the backend that the base `shoes` script can
run to start applications.

Leveraging this new scheme, the shoes-swt gem now installs a shoes-swt
executable, along with the necessary code to generate a compatible
`shoes-backend`. Still more to do making that generation fully
cross-platform for SWT, but it's a start.

shoes-picker is the heart of selecting a backend from the available
gems. Right now this only supports a single backend, but this will be
easy to extend to allow selecting different backends.