Default left and top to nil
This should resolve #948.

When creating a shape with the dsl, you can:

 1. specify a left and top
 2. specify a left and top in the styles hash
 3. forgo setting the left and top

In the first case the left and top will be written into the styles hash
before it is passed along to `Shoes::Shape`.

In the third case the left and top values will be nil.

`Shoes::Shape` no longer accepts a left and a top as arguments.