Fix for overlapping text
When flowing text to a second layout, we would stop short of consuming
all the vertical space provided us, but then position the second layout
based on the height of the first layout. Sometimes this resulted in
overlapping text.

Instead, we will now keep putting text into into the first layout until
it completely fills the height--even if it goes over by a little--so
that we get a smoother flow into the second layout.