Don't calculate element dimensions unnecessarily
During painting with fill and stroke we were calculating the left, top,
width and height. Because these are calculated properties and are called
really often during painting, it's actually a slow bit, despite the
dimensions only being used from gradients.

This commit changes so we pass along the DSL element that things are
being applied to, and only do the calculation for gradients, so simple
colors and patterns are faster.