Convert samples/simple-form.rb to use images
Previously samples/simple-form.rb used backgrounds for its image-based
borders. The nice part about this was that it made the left and right
edges repeating, so if the window was dragged larger, they'd carry on
down the edges as far as desired.

The problem, though, is that SWT doesn't allow for choosing where a
repeating image pattern begins, so the image on the right is offset
incorrectly. This isn't fixable without a huge amount of code that
doesn't even work properly for all art elements and situations (see #1105
for an attempt and where it fell apart)

All that to say, just use images for the borders.