Gentler, kinder error handling for downloads
Changes to percolate errors during Download up to the app were making it
so a simple 404 caused a crash. Oops, not quite what we want.

But we don't want things to go away silently either. In light of that,
we now log when there's a downloading failure, and also provide a
callback proc that can be set to handle errors on the Download yourself
if you want it.

This also prompted some revision of the tests, which were being
unnecessarily inexact in their expectations around which proc would get
executed at what point in time.