Clip when drawing to graphics context based on parent
This adds a method to the poorly named Shoes::Swt::Common::Resource
class (which is really all about graphics context management and should
get a refresh there) to clip the region drawn to a parent's bounds for
the duration of a block. Normal drawing is restored afterward.

Applying this to TextBlock painting will properly cut off display of
text when it flows outside the slot's boundaries.