Don't advance position for right aligned text
As noted in https://github.com/shoes/shoes4/issues/1194 we get odd
effects in flowing if we allow right-aligned elements bump the current
position when doing slot alignment.

While this is a breaking change with Shoes 3, it behaves more
consistently, and we'll be happier with it in the long haul this way.