Fix several calculator issues
In any physical, real-world calculator I can recall using, if I wanted to compute 2+2 and 3+3 I would click the following buttons:

2 + 2 = 3 + 3 =

But if I do that in this program,

2 + 2 =

shows "4", as expected, but then clicking 3 changes the display to "43".

I appreciate that the sample is intended as an example of using Shoes, not an example of how to write calculator logic.  But this problem makes it hard to actually use the calculator.

Another fix is that on most calculators, doing:

1 + 2 + 3 + 4 + 5 + 6 + 7 + 8 + 9 =

shows the running total every time + is clicked.  This program did not display any result until = was clicked.

Renamed several instance variable to be more precise.

Corrected 4 RuboCop issues.